+++ examples/resources/superset_user/import.sh
@@ -1,0 +1,2 @@
+# User can be imported by specifying the numeric identifier of the User id
+terraform import superset_user.example_user 213